+++ Elements/src/ContentElement.cs
@@ -24,9 +24,28 @@ namespace Elements
         /// <param name="name">The name of this element.</param>
         /// <param name="additionalProperties">The string json serialization of a dictionary of additional parameters.</param>
         public ContentElement(string @gltfLocation, BBox3 @boundingBox, double @gltfScaleToMeters, Vector3 @sourceDirection, Transform @transform, Material @material, Representation @representation, bool @isElementDefinition, System.Guid @id, string @name, string @additionalProperties)
-        : this(@gltfLocation, @boundingBox, @gltfScaleToMeters, @sourceDirection, @transform, @material, @representation, @isElementDefinition, @id, @name)
+        : this(@gltfLocation, @boundingBox, @gltfScaleToMeters, @sourceDirection, null, @transform, @material, @representation, @isElementDefinition, @id, @name)
         {
             this.AdditionalProperties = Newtonsoft.Json.JsonConvert.DeserializeObject<Dictionary<string, object>>(@additionalProperties);
+        }
+
+        /// <summary>
+        /// This constructor adds backwards compatibility with the old constructor, before Symbols were added.
+        /// </summary>
+        /// <param name="gltfLocation">The path to the .glb file.</param>
+        /// <param name="boundingBox">The BBox3 of this Content Element.</param>
+        /// <param name="gltfScaleToMeters">The number required to scale this contents dimensions to meters.  Used during gltf merging.</param>
+        /// <param name="sourceDirection">The direction the element was facing when it was extracted from it's source.</param>
+        /// <param name="transform">The transform of this ContentElement.</param>
+        /// <param name="material">The material, used for the BBox representation of this element.</param>
+        /// <param name="representation">The representation which will be updated when needed.</param>
+        /// <param name="isElementDefinition">Should the element be used to create instances, or should it be inserted into a 3D scene.</param>
+        /// <param name="id">The guid of this element.</param>
+        /// <param name="name">The name of this element.</param>
+        public ContentElement(string @gltfLocation, BBox3 @boundingBox, double @gltfScaleToMeters, Vector3 @sourceDirection, Transform @transform, Material @material, Representation @representation, bool @isElementDefinition, System.Guid @id, string @name)
+        : this(@gltfLocation, @boundingBox, @gltfScaleToMeters, @sourceDirection, null, @transform, @material, @representation, @isElementDefinition, @id, @name)
+        {
+
         }