+++ Elements/src/Geometry/Representation.cs
@@ -23,5 +23,13 @@ namespace Elements.Geometry
         {
             return new Representation(solidOperation);
         }
+
+        /// <summary>
+        /// A flag to disable CSG operations on this representation. Instead,
+        /// all solids will be meshed, and all voids will be ignored.
+        /// </summary>
+        [Newtonsoft.Json.JsonIgnore]
+        [System.ComponentModel.EditorBrowsable(System.ComponentModel.EditorBrowsableState.Never)]
+        public bool SkipCSGUnion { get; set; } = false;
     }
 }