+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 0.9.3
+
+### Changed
+
+- Added `Symbols` property to `ContentElement`.
+- Introduce a `SkipCSGUnion` flag on Representation, as a hack to get around CSG failures.
+
 ## 0.9.2
 
 ### Added
@@ -25,6 +32,7 @@
   - `Polygon.UnionAll(params Polygon[] polygons)`
   - `Polygon.Difference(params Polygon[] polygons)`
   - `Polygon.Union(params Polygon[] polygons)`
+- `Profile.Offset()`
 
 ### Changed